feat: 블록 에디터 모듈화: 업데이트 콘텐츠 모듈화
